task/requirements.txt
feedparser==6.0.11
asyncpg==0.29.0
aiohttp==3.9.5
aiogoogle==5.12.0
SQLAlchemy==2.0.31
python-dateutil==2.9.0.post0
pydantic==2.8.2
pytz==2024.1